cifsd: Use netdevice_notifier to configure TCP listeners
This commit changes how TCP listeners are initialized.
ksmbd_tcp_init() registers a notifier for netdevices state changes.
As a result, TCP listeners are set up whenever a NETDEV_UP event is
received for an interface.

Since 'UP' events are replayed on registration, listeners are initialized
for all configured devices. ksmbd_tcp_set_interfaces() allocates an
interface structure regardless of the netdevice's state. This lets us
create a listener when the interface comes up.

Similarly, 'DOWN' events are received on unregistration which means that
the kthread/socket cleanup is always done in ksmbd_netdev_event().

Signed-off-by: Marios Makassikis <mmakassikis@freebox.fr>
Signed-off-by: Namjae Jeon <namjae.jeon@samsung.com>