cifsd: fix the issue of change the directory to the file
Nautilus send smb2 setinfo request with ATTR_NORMAL attribute when
directory is copied into ksmbd share. So ksmbd set it to file attribute.
It causes the issue of making a directory appear as a file by ksmbd
setting it to fp->f_ci->m_fattr. This patch add the check to avoid setting
ATTR_NORMAL_LE flags to directory.

Signed-off-by: Namjae Jeon <namjae.jeon@samsung.com>