cifs: deprecate 'enable_negotiate_signing' module param
Deprecate enable_negotiate_signing module parameter as it's irrelevant
since the requested dialect and server support will dictate which
algorithm will actually be used (*).

Send a negotiate signing context on every SMB 3.1.1 negotiation now.
The difference with the current behaviour is that AES-GMAC is set as the
preferred algorithm in the context (done in the AES-GMAC introduction
commit).

AES-CMAC will still be used instead, if:
  - SMB 3.0.x negotiated
  - SMB 3.1.1 negotiated, but server doesn't support AES-GMAC

IOW, this is a safe change.

Warn the user if the module was loaded with 'enable_negotiate_signing=0'.

Signed-off-by: Enzo Matsumiya <ematsumiya@suse.de>